rename AudioFile/{AIFCCompression => aifc_compression}.hpp
