+++ include/AudioFile/aifc_compression.hpp
